requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
openai==0.27.8
Pillow==9.5.0
pybit==5.5.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
tqdm==4.65.1
urllib3==2.0.4
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
